replace the actually used machine definitions to build layer information
As the assertions show that they are equal for the existing tests, we can replace the existing logic
to translate the old definitions and directly use the new ones.

Also removes debug print statements and replaces the use of num_of_hyperthreading_siblings, so we can use the one
in the machine definition (which we later can easily use with pytest parameterization).