move spreading_memory_region_key to auxillery function
To reduce the complexity of core_allocation_algorithm, it seems like a good idea to decouple the
high level parts of the algorithm itself from the actual crawling of lists and dicts.

Note that this commit only duplicates the code, to be able to use an assertion to make sure the new
auxillery function returns equivalent results. The existing code block (and assertion) is removed in
the next commit.