Detect and forbid duplicate run-definition names
If run-definition names are duplicated, BenchExec writes their results
in files with the same name, so it overwrites results.
This never makes sense and it is better UX if we detect and forbid this.

With this commit we add checks we forbid only as few as cases as possible:
We still allow duplicate run-definition names and even more than one
unnamed run definition as long as at most run definition for each name
is actually executed
(one can achieve this for example by having different <tasks> names
for each of them and filtering based on this).
Furthermore, we allow duplicate run-definition names if each of the
run definitions has a single <tasks> tag with a unique name
(because then BenchExec already uses a unique run-definition name
that includes the <tasks> name).
So we forbid only cases where actual overwriting of result files would occur.

For more predictable BenchExec behavior it might even be better to
forbid more cases (e.g., simply require that all run-definition names in
a file need to be unique), but we should do such a change only for a
major revision.

The only case where this change can break existing workflows is
if there are run definitions with duplicate names
but each as more than one <tasks> tags, all with unique names,
and if the user was ignoring the result file for the run definition
and using only the result files per task set.
But it is still too error-prone to allow a case where BenchExec
overwrites its own files.