Prevent overwriting of result files per <tasks> tag
This could happen if several <tasks> tags for a run definition have the
same non-empty name. This is of course error prone and we want to avoid it.
Because many users ignore the result files per task set probably anyway,
we do not yet forbid this, because it might be a breaking change.
But we stop writing output files for task sets with non-unique names
and show a warning.

Note that if there are <tasks> tags without a name,
BenchExec internally assigns them their index as name,
so these are not affected.