Provide command-line arguments for choosing result files of BenchExec
BenchExec supports writing the benchmark results into several XML files.
So far, it produces one XML file for each <rundefinition> in the benchmark
definition, and another XML file for each <tasks> tag for each run definition.
The names of the files have the names of the run definitions and task
sets as part of their suffix.
An exception is that if there is a run definition with a single task set,
there would be two result files with the exact same set of results in them,
so BenchExec creates only one of them.

Writing out the results in two files (per run definition and per task set)
can be convenient, but is often also redundant.
Furthermore, in the mentioned exception the name of the produced result
file may not match the expectation of the user (it includes the names of
both run definition and task set).

To solve this, we introduce two new command-line arguments:
- --results-per-rundefinition
- --results-per-taskset

If the former is given, BenchExec writes a result file per run
definition. If the latter is given, BenchExec writes a result file per
task set (for each run definition). If both are given, BenchExec writes
both kinds of result files (even if some of them end up being identical).
This is clear and consistent.

If none of the arguments is given, the behavior of BenchExec is as
before for now. In the next major version of BenchExec we will change that
such that we get consistent output by default.

Furthermore, some potentially confusing names of result files
happen due to the fact that the names of run definitions and task sets
are optional, and that named and unnamed run definitions and task sets
can even be used together in the same benchmark definition.
We change this now and add the following rules:
- Run definitions need to have non-empty unique names
  if any of the new arguments is given.
  Exception: If there is a single run definition, its name is optional.
- Task sets need to have non-empty unique names
  if --results_per_taskset is given.
This allows both simple benchmark definitions with little boilerplate
and no names and complex benchmark definitions with many tags
while still ensuring predictable names of result files.
We can assume that most benchmark definitions of users
will already conform to these rules.
If none of the new arguments is given, the behavior remains as before
for now.

This implements the core part of #983.
What is not done yet:
- The default behavior of BenchExec (if none of the new arguments is given)
  is the same as before. This will change in the next major version.
- If only --results_per_taskset is given, no intermediate files with
  partial results during benchmarking are written. This is a TODO.
- If only --results_per_taskset is given, the hint at the end of
  BenchExec's execution on how to produce tables is not printed.
  This is a TODO, but depends on #977.
- Tests for the new behavior.
- Documentation and examples in the doc/ directory.