Use Ruff as code formatter for Python code instead of Black
Ruff is a newer tool that shares the same spirit
and has the goal of producing basically the same format as Black.
There are only minimal differences, but where there are,
Ruff is slightly nicer (cf. diff of this commit).
It is also faster and we already use it as linter,
so developers have to worry about one tool less.