Add tests for set of generated result files of benchexec
Part of #983.

Depending on which tags in the benchmark definition have names
the set of result files that benchexec generates changes.
This can be confusing and we want to improve it.
As first step add tests for the current behavior
such that we do not accidentally change it
when adding new command-line options with different behavior.